Added a debug feature for toggling environments
