+++ vimbisopay_app/.gitignore
@@ -70,6 +70,7 @@ app.*.map.json
 
 # Sensitive configuration files
 .env
+.api_env_prod
 lib/core/config/api_config.dart
 lib/core/config/encryption_config.dart
 scripts/github_config.sh
