+++ vimbisopay_app/lib/main.dart
@@ -8,6 +8,7 @@ import 'dart:async';
 import 'package:lottie/lottie.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:vimbisopay_app/core/config/api_config.dart';
 import 'package:vimbisopay_app/presentation/blocs/notifications/notifications_bloc.dart';
 import 'package:vimbisopay_app/presentation/screens/intro_screen.dart';
 import 'package:vimbisopay_app/presentation/screens/create_account_screen.dart';
@@ -122,6 +123,13 @@ void main() async {
     print('Loading environment variables...');
     await dotenv.load(fileName: '.env');
     print('Environment variables loaded successfully');
+    
+    // Initialize API configuration
+    print('Initializing API configuration...');
+    await ApiConfig.initialize();
+    await ApiConfig.refreshEnvironment();
+    print('API configuration initialized successfully');
+    Logger.data('API environment: ${ApiConfig.environmentName}');
 
     // Initialize Firebase first
     print('Initializing Firebase...');
